Create feature directories and move slice into their respective directories
